test/regression/reftest/vendor/webkit/fast/dom/node-legacy-event-listener.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/wpt-test-pre.js"></script>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Test addEventListener() and removeEventListener() fail silently if arguments are missing.");

shouldThrow('document.addEventListener("foo")');
shouldThrow('document.removeEventListener("bar")');

shouldThrow('document.addEventListener()');
shouldThrow('document.removeEventListener()');

</script>
<script src="../../resources/js-test-post.js"></script>
<script src="../../resources/wpt-test-post.js"></script>
</body>
</html>
